+++ components/restreamer/CHANGELOG.md
@@ -20,11 +20,14 @@ All user visible changes to this project will be documented in this file. This p
 ### Added
 - Dashboard:
   - Remove all clients ([#386])
-  - Star/Stop playing file ([#386])
+  - Star/Stop playing file by file name prefix ([#415])
 - Playlist
   - Status component ([#343], [#348])
   - Make playlist optional on import ([#405])
-
+  - Add single file from Google Drive to playlist ([#415])
+  - Show notification icon if there are some files not having stream info ([#415])
+  - Compare stream information of files in a playlist and display an icon if they are not equal ([#415])
+  - Compare stream information of endpoints with the stream info of the currently playing file from the playlist ([#415])
 ### Miscellaneous
 - Update [FFmpeg] to 6.0  ([#375]);
 
@@ -38,6 +41,7 @@ All user visible changes to this project will be documented in this file. This p
 [#375]: /../../pull/375
 [#386]: /../../pull/386
 [#405]: /../../pull/405
+[#415]: /../../pull/415